+++ src/PaymentData.php
@@ -13,7 +13,7 @@ use Pronamic\WordPress\Pay\Payments\Items;
  * Company: Pronamic
  *
  * @author  Remco Tolsma
- * @version 2.1.0
+ * @version 2.1.2
  * @since   1.0.0
  */
 class PaymentData extends Pay_PaymentData {
@@ -88,7 +88,7 @@ class PaymentData extends Pay_PaymentData {
 		$item->set_price( $this->data['total_cost'] );
 		$item->set_quantity( 1 );
 
-		$items->addItem( $item );
+		$items->add_item( $item );
 
 		return $items;
 	}