+++ CHANGELOG.md
@@ -7,6 +7,11 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 -
 
+## [2.1.2] - 2019-12-22
+- Improved error handling with exceptions.
+- Updated usage of deprecated `addItem()` method.
+- Updated payment status class name.
+
 ## [2.1.1] - 2019-08-26
 - Updated packages.
 
@@ -29,7 +34,8 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## 1.0.0 - 2015-01-20
 - First release.
 
-[unreleased]: https://github.com/wp-pay-extensions/event-espresso-legacy/compare/2.1.1...HEAD
+[unreleased]: https://github.com/wp-pay-extensions/event-espresso-legacy/compare/2.1.2...HEAD
+[2.1.2]: https://github.com/wp-pay-extensions/event-espresso-legacy/compare/2.1.1...2.1.2
 [2.1.1]: https://github.com/wp-pay-extensions/event-espresso-legacy/compare/2.1.0...2.1.1
 [2.1.0]: https://github.com/wp-pay-extensions/event-espresso-legacy/compare/2.0.0...2.1.0
 [2.0.0]: https://github.com/wp-pay-extensions/event-espresso-legacy/compare/1.0.3...2.0.0